LayoutTests/http/tests/navigation/page-cache-getUserMedia-pending-promise.html

<!-- webkit-test-runner [ UsesBackForwardCache=true ] -->
<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description('Tests that a page with a pending getUserMedia request is able to go into the back/forward cache.');
jsTestIsAsync = true;
let restoredFromPageCache = false;
let getUserMediaRejected = false;

if (window.testRunner)
    testRunner.setUserMediaPermission(true);

function getUserMediaCompleted() {
    // Avoid invoking getUserMedia again.
    window.removeEventListener("pagehide", onPageHide);
    shouldBeTrue("restoredFromPageCache");
    // Pending request gets denied on navigation.
    shouldBeTrue("getUserMediaRejected");
    finishJSTest();
}

function onPageHide(event) {
    debug("pagehide - " + (event.persisted ? "" : "not ") + "entering cache");
    if (!event.persisted) {
        testFailed("Page failed to enter the Page Cache");
        finishJSTest();
    }

    navigator.mediaDevices.getUserMedia({ audio: true }).then(getUserMediaCompleted, () => {
        getUserMediaRejected = true;
        getUserMediaCompleted();
    });
}

function onPageShow(event) {
    debug("pageshow - " + (event.persisted ? "" : "not ") + "from cache");
    if (event.persisted) {
        testPassed("Page was restored from Page Cache");
        restoredFromPageCache = true;
    }
}

window.addEventListener("pageshow", onPageShow);
window.addEventListener("pagehide", onPageHide);

onload = () => {
    setTimeout(() => {
        testLink.click();
    }, 0);
}
</script>
<a id="testLink" href="resources/page-cache-helper.html" style="display: none">Link</a>
</body>
</html>
